refactor: extract the columnar master write path out of write.rs
Merging four branches pushed write.rs to 2545 lines, past its 2500
ceiling. Project doctrine is to split a file that outgrows the ceiling
rather than raise the allowlist, so the three items that reach the
per-type master Arc<ColumnStore> move to columnar_write.rs -- a real
seam, not an arbitrary cut: execute_set and execute_remove must agree
on all three, since a property removed on a node's forked store while
the master keeps the value is resurrected by the next clause's sweep.

execute_set's function exceptions tightened downward to the measured
508->479 lines and 70->64 branches by hand, not --refresh-functions.
